+++ src/client/Microsoft.Identity.Client/Internal/Requests/Interactive/InteractiveRequest.cs
@@ -9,7 +9,9 @@ using Microsoft.Identity.Client.Core;
 using Microsoft.Identity.Client.Instance;
 using Microsoft.Identity.Client.Internal.Broker;
 using Microsoft.Identity.Client.OAuth2;
+using Microsoft.Identity.Client.PlatformsCommon.Shared;
 using Microsoft.Identity.Client.UI;
+using Microsoft.Identity.Client.Utils;
 
 namespace Microsoft.Identity.Client.Internal.Requests
 {
@@ -63,8 +65,48 @@ namespace Microsoft.Identity.Client.Internal.Requests
         {
             await ResolveAuthorityAsync().ConfigureAwait(false);
             cancellationToken.ThrowIfCancellationRequested();
-            MsalTokenResponse tokenResponse = await GetTokenResponseAsync(cancellationToken)
-                .ConfigureAwait(false);
+            MsalTokenResponse tokenResponse = null;
+            if (ServiceBundle.Config.IsBrokerEnabled && DesktopOsHelper.IsMacConsoleApp())
+            {
+                var macMainThreadScheduler = MacMainThreadScheduler.Instance();
+                if (!macMainThreadScheduler.IsCurrentlyOnMainThread())
+                {
+                    throw new MsalClientException(
+                        MsalError.WamUiThread,
+                        "Interactive requests with mac broker enabled must be executed on the main thread on macOS.");
+                }
+                bool messageLoopStarted = macMainThreadScheduler.IsRunning();
+                var tcs = new TaskCompletionSource<MsalTokenResponse>();
+                _ = Task.Run(async () =>
+                {
+                    try
+                    {
+                        MsalTokenResponse response = await GetTokenResponseAsync(cancellationToken).ConfigureAwait(false);
+                        tcs.SetResult(response);
+                    }
+                    catch (Exception ex)
+                    {
+                        _logger.Error($"Error in background GetTokenResponseAsync: {ex}");
+                        tcs.SetException(ex);
+                    }
+                    finally
+                    {
+                        if (!messageLoopStarted)
+                            macMainThreadScheduler.Stop();
+                    }
+                });
+                if (!messageLoopStarted)
+                {
+                    _logger?.Verbose(() => "Mac broker console app scenario needs to start a message loop internally.");
+                    macMainThreadScheduler.StartMessageLoop();
+                }
+                tokenResponse = await tcs.Task.ConfigureAwait(false);
+            }
+            else
+            {
+                tokenResponse = await GetTokenResponseAsync(cancellationToken)
+                    .ConfigureAwait(false);
+            }
             return await CacheTokenResponseAndCreateAuthenticationResultAsync(tokenResponse)
                 .ConfigureAwait(false);
         }